make `LazyString` immutable
Currently a function that constructs `LazyString` but isn't decided to
throw isn't concrete-foldable because `LazyString` is defined as mutable
type and the effect analysis conservatively taints `:consistent`-cy on
a construction of mutable object:
```julia
julia> let eff = Base.infer_effects((Int,)) do a
               if a < 0
                   throw(DomainError(lazy"$a isn't positive"))
               end
               return a
           end
           display(eff)
           Core.Compiler.is_foldable(eff)
       end
(?c,+e,?n,+t,+s)
false
```

This commit makes `LazyString` immutable and let it re-construct actual
string representation on every access. This of course may lead to
worse performance when we print the same `LazyString` object again and
again, but presumably such an case would be very rare given that
`LazyString` is supposed to be mainly used on error path.